+++ tests/TestUtils/ExtraUtilsTests.php
@@ -7,6 +7,7 @@ use PHPUnit\Framework\TestCase;
 use Raptor\TestUtils\Exceptions\BadMethodException;
 use Raptor\TestUtils\ExtraUtils;
 use RaptorTests\TestUtils\Utils\InvokeMethodTestObject;
+use RuntimeException;
 
 /**
  * @author Mikhail Kamorin aka raptor_MVK
@@ -62,6 +63,33 @@ final class ExtraUtilsTests extends TestCase
         return [
             'protected' => ['reverse', [$palindrome . $extra], $extra . $palindrome],
             'private' => ['repeat', [$palindrome, $count], str_repeat($palindrome, $count)]
+        ];
+    }
+
+    /**
+     * Checks that method _expectExceptionExactMessage_ accepts correct message.
+     *
+     * @param string $message
+     *
+     * @dataProvider exactMessageDataProvider
+     */
+    public function testExpectExceptionExactMessage(string $message): void
+    {
+        $this->expectExceptionExactMessage($message);
+
+        throw new RuntimeException($message);
+    }
+
+    /**
+     * Provides test data for testing method _expectExceptionExactMessage_.
+     *
+     * @return array [ [ message ], ... ]
+     */
+    public function exactMessageDataProvider(): array
+    {
+        return [
+            'simple' => ['Error occurred'],
+            'escaped characters' => ['\+*?[^]$(){}=!<>|:-#/']
         ];
     }
 }